Delete originscrip, rename new script,adjust yaml.
