+++ .github/workflows/break-sentence.yml
@@ -37,7 +37,7 @@ jobs:
       - name: Download packages
         run: pip install requests openpyxl click pandas
       - name: Execuate break-sentence scripts.
-        run: python ./BreakSentence/breakSentence_temp.py  --name ${{ inputs.name }} --type ${{ inputs.type }} --lang ${{ inputs.lang }} run
+        run: python ./BreakSentence/breakSentence.py  --name ${{ inputs.name }} --type ${{ inputs.type }} --lang ${{ inputs.lang }} run
         env:
             AZURE_API_KEY1 : ${{ secrets.AZURE_API_KEY1}}
             AZURE_LOCATION: ${{ secrets.AZURE_LOCATION }}
